requirements.txt
imageio[ffmpeg]
streamlit
pillow